Lots of changes in this one
Adds CSV import for both file types, reads the files off of local
storage on SD cards, and adds stuff to the database. Still needs to be
tested with actual device for SD card things and needs to be smarter
about not adding duplicates to the database. Uncomment and run
'testCSVImport()' in LoginActivity to look for files on the SD card and
add them to the database.